CB-3638 Delete service-managed dbs when server deleted
When a service-managed database server is deleted (as part of
deregistration after termination), the databases that were created by
redbeams on it are deleted first. Otherwise, deletion, which is actually
archival, happens as before.

Both DatabaseConfig and DatabaseServerConfig now have implementations of
the unsetRelationsToEntitiesToBeDeleted which clear out relations from
those entities to each other as part of the deletion (archival) process.

For simplicity, the DatabaseConfigService::deleteOne method is renamed
to simply delete, overriding but calling to its superclass
implementation. The full implementation of the method accepts a
skipDeletionOnServer flag, so that database deletion due to a database
server deletion does not try to contact the deleted server; at that
point, the database is already gone. The method also accepts a force
flag so that failure to delete a database on a server does not have to
fail the deletion attempt overall. (This mechanism isn't directly used
yet.)

Other changes:

* DatabaseV4Response gains a new resourceStatus field, for more
  information in the API.
* DatabaseConfigTest and DatabaseV4ResponseTest are added.
* DatabaseServerV4ResponseTest is slightly expanded.
* The unused method DatabaseConfigService::archive is removed.

+++ redbeams/src/main/java/com/sequenceiq/redbeams/service/dbconfig/DatabaseConfigService.java
@@ -108,11 +108,6 @@ public class DatabaseConfigService extends AbstractArchivistService<DatabaseConf
         }
     }
 
-    public void archive(DatabaseConfig databaseConfig) {
-        databaseConfig.setArchived(true);
-        repository.save(databaseConfig);
-    }
-
     public DatabaseConfig getByCrn(String resourceCrn) {
         Crn crn = Crn.safeFromString(resourceCrn);
         Optional<DatabaseConfig> resourceOpt = repository.findByResourceCrn(crn);
@@ -150,18 +145,18 @@ public class DatabaseConfigService extends AbstractArchivistService<DatabaseConf
                     String.format("Database(s) not found: %s", String.join(", ", notFoundDatabaseConfigs)));
         }
         return foundDatabaseConfigs.stream()
-                .map(this::deleteOne)
+                .map(this::delete)
                 .collect(Collectors.toSet());
     }
 
     public DatabaseConfig deleteByCrn(String resourceCrn) {
         DatabaseConfig resource = getByCrn(resourceCrn);
-        return deleteOne(resource);
+        return delete(resource);
     }
 
     public DatabaseConfig deleteByName(String name, String environmentCrn) {
         DatabaseConfig resource = getByName(name, environmentCrn);
-        return deleteOne(resource);
+        return delete(resource);
     }
 
     public String testConnection(String databaseConfigName, String environmentCrn) {
@@ -180,17 +175,46 @@ public class DatabaseConfigService extends AbstractArchivistService<DatabaseConf
                 .collect(Collectors.joining("; "));
     }
 
-    private DatabaseConfig deleteOne(DatabaseConfig databaseConfig) {
+    @Override
+    public DatabaseConfig delete(DatabaseConfig databaseConfig) {
+        return delete(databaseConfig, false, false);
+    }
+
+    /**
+     * Deletes a database configuration. If the database is service-managed,
+     * then also deletes the database from its home server unless
+     * <code>skipDeletionOnServer</code> is true. If deletion from the server
+     * fails, deletion fails unless <code>force</code> is true. (Currently,
+     * deletion from the server is the only action that <code>force</code>
+     * applies to.)
+     *
+     * @param  databaseConfig       database to delete
+     * @param  force                whether to force deletion
+     * @param  skipDeletionOnServer whether to skip deleting the database on its server
+     * @return                      deleted database
+     */
+    public DatabaseConfig delete(DatabaseConfig databaseConfig, boolean force, boolean skipDeletionOnServer) {
         LOGGER.info("Deleting database with name: {}", databaseConfig.getName());
 
-        DatabaseConfig deletedConfig = delete(databaseConfig);
-
-        // Only delete database from server if successfully removed from redbeams
         if (databaseConfig.getStatus() == ResourceStatus.SERVICE_MANAGED) {
-            deleteServiceManagedDatabase(databaseConfig);
+            if (!skipDeletionOnServer) {
+                // Only delete database from redbeams once successfully removed from server
+                try {
+                    deleteServiceManagedDatabase(databaseConfig);
+                } catch (RuntimeException e) {
+                    if (force) {
+                        LOGGER.warn("Deletion for database '{}' failed, continuing because termination is forced",
+                                databaseConfig.getName(), e);
+                    } else {
+                        throw e;
+                    }
+                }
+            } else {
+                LOGGER.debug("Skipping deletion on server");
+            }
         }
 
-        return deletedConfig;
+        return super.delete(databaseConfig);
     }
 
     private void deleteServiceManagedDatabase(DatabaseConfig databaseConfig) {
